Rename dfs package to msdfsc package.
