+++ src/main/java/com/hierynomus/msdfsc/ReferralCache.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.hierynomus.mssmb2.dfs;
+package com.hierynomus.msdfsc;
 
 import static com.hierynomus.protocol.commons.EnumWithValue.EnumUtils.isSet;
 
@@ -21,7 +21,7 @@ import java.util.ArrayList;
 import java.util.Hashtable;
 import java.util.List;
 
-import com.hierynomus.mssmb2.dfs.SMB2GetDFSReferralResponse.ReferralHeaderFlags;
+import com.hierynomus.msdfsc.SMB2GetDFSReferralResponse.ReferralHeaderFlags;
 
 /**
  * [MS-DFSC].pdf: 3.1.1 Abstract Data Model
