+++ src/main/java/com/hierynomus/msdfsc/UserHostSessionTable.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.hierynomus.mssmb2.dfs;
+package com.hierynomus.msdfsc;
 
 import java.util.Hashtable;
